Pre-process <imagedata> tags
DocBook XML sources need to have a width attribute set up in <imagedata> tags
to specify the optimal size for images in print media. When dealing with the
web, however, it's best to let Bootstrap figure out how to render responsive
images.

Why this change?

Not setting a width (or using a 100% width) in the source file makes its
processing from XML to FO (the necessary intermediate step before rendering
as PDF) extremely slow, presumably because the formatting engine has to
reflow the entire book to figure out what "100%" means in that context. When
using a fixed value, such as 14cm, the processing is much faster.

However, if we change the documentation source to optimize for print with
fixed size the web version gets really small images as a result. That's
because the width attribute of the <imagedata> tag is transformed by the XSL
stylesheets into an equivalent pixel value width for the resulting <image>
tag of the HTML output. For 14cm this ends up being some silly small value
around 400px which makes the images thumbnail-sized, therefore unreadable.

So the best way to fix both issues is to spend less than a second for pre-
processing the <imagedata> tags in DocImport, removing the width, letting the
template's Bootstrap handle the responsive sizes of the images.