Workaround for broken PHP include boolean evaluation
One of our clients had a PHP 7.3 server which did not work
properly with `if(!@include(…) && !defined(…))` constructs,
but only when it came to FEF's version includes. We have no
good theory why so we decided to change all version file
includes to not use this construct out of an abundance of
caution.